backend/requirements.txt
fastapi
uvicorn
python-dotenv
valkey
blake3
pytest 
httpx